+++ src/util/mod.rs
@@ -6,10 +6,10 @@ pub use self::error::*;
 pub use self::lock::*;
 
 #[inline]
-pub const fn align_down(val: u64, align: u64) -> u64 {
+pub const fn align_down(val: usize, align: usize) -> usize {
     val / align * align
 }
 #[inline]
-pub const fn align_up(val: u64, align: u64) -> u64 {
+pub const fn align_up(val: usize, align: usize) -> usize {
     (val + align - 1) / align * align
 }
